serverPy/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
marshmallow==3.20.1
marshmallow-sqlalchemy==1.1.0
